+++ contrib/local_azure_tests.py
@@ -14,12 +14,14 @@ commands = [
 if os.getenv("LOCAL_DEV"):
     additional_commands = [
         "electrumsv-sdk install node",
-        "electrumsv-sdk install electrumx",
+        "electrumsv-sdk install simple_indexer",
+        "electrumsv-sdk install reference_server",
         "electrumsv-sdk install electrumsv",
         "electrumsv-sdk install whatsonchain",
         "electrumsv-sdk start status_monitor",
         "electrumsv-sdk start node",
-        "electrumsv-sdk start electrumx",
+        "electrumsv-sdk start simple_indexer",
+        "electrumsv-sdk start reference_server",
         "electrumsv-sdk start electrumsv",
         "electrumsv-sdk node generate 1",
         "electrumsv-sdk start whatsonchain",
@@ -28,7 +30,8 @@ else:
     additional_commands = [
         "electrumsv-sdk start --background status_monitor",
         "electrumsv-sdk start --background node",
-        "electrumsv-sdk start --background electrumx",
+        "electrumsv-sdk start --background simple_indexer",
+        "electrumsv-sdk start --background reference_server",
         "electrumsv-sdk start --background electrumsv",
         "electrumsv-sdk node generate 1",
         "electrumsv-sdk start --background whatsonchain",
@@ -36,12 +39,12 @@ else:
 
 and_more_commands = [
     "electrumsv-sdk stop node",
-    "electrumsv-sdk stop electrumx",
+    "electrumsv-sdk stop simple_indexer",
+    "electrumsv-sdk stop reference_server",
     "electrumsv-sdk stop electrumsv",
     "electrumsv-sdk stop whatsonchain",
 
     "electrumsv-sdk reset --id=node1",
-    "electrumsv-sdk reset --id=electrumx1",
     "electrumsv-sdk reset --id=electrumsv1",
     f"{sys.executable} -m pylint --rcfile ../.pylintrc {Path(MODULE_DIR).parent.joinpath('electrumsv_sdk')}",
 ]