+++ electrumsv_sdk/builtin_components/electrumx/__init__.py
@@ -1,1 +1,0 @@
-from .electrumx import Plugin