+++ electrumsv_sdk/builtin_components/peer_channels/__init__.py
@@ -1,1 +1,0 @@
-from .peer_channels import Plugin