+++ electrumsv_sdk/commands.py
@@ -31,7 +31,7 @@ def install(component_type: str, repo: str = "", branch: str = "",
 
 
 def _validate_network(network: str, component_type: str) -> None:
-    components_with_network_option = {'node', 'electrumx', 'electrumsv'}
+    components_with_network_option = {'node', 'electrumsv'}
     if network != '' and component_type not in components_with_network_option:
         raise ValueError(f"Can only specify 'network' for {components_with_network_option}")